refactor(game): change score type from float to int
Change GameState score field from float to int.
Remove unnecessary (int) cast in renderer.c.
Add explicit cast in game.c for score increment.
Improves type consistency since score is displayed as integer.

+++ include/game.h
@@ -13,7 +13,7 @@ typedef struct {
   GameStatus status;
   Player player;
   Pipe pipes[MAX_PIPES];
-  float score;
+  int score;
   bool shouldQuit;
 } GameState;
